requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5
asyncpg==0.29.0 